Support l3 dvr patch ports
Support of l3 dvr by creating logical router port on lrouter and
its patch port on connected lswitch when router interface is being
added. ovn_api is also updated according to latest OVN NB schema
for this purpose.

With this support, ports on different subnets are able to talk to
each other by connecting the subnets to the same router.

Closes-Bug: #1514649
Change-Id: I0541b258cc357e2a151b24d865feef181c85056d