WIP: Serialize conflicting requests in task queue
In ML2 mechanism driver, OVN updates are performed in post-commit,
which is after neutron DB transaction. With multi-workers/nodes,
the order of OVN updates are not guaranteed, which means the order
of OVN updates can be different from that of DB transaction commit
when there are racings between workers/nodes. This will result in
inconsistent data between OVN and Neutron DB.

This patch fixes the problem by serializing conflicting requests
in a task queue in DB. A place holder is marked in the queue in
precommit within the same DB transaction of the neutron DB change,
and during postcommit it checks and waits if there are conflicting
changes pending in the queue. This ensures ordering while keeping
the sync behaviour unchanged from client point of view, i.e. it
doesn't return until ovn nb operation complete.

TODO:
- Now only port-update is involved, other operations TBD
- Indirect dependency mentioned by bug #1611852 will be fixed in
a separate patch

Closes-bug: #1605089
Change-Id: If1a3a9e1841b7539a678a01182985cacb19d9dcc